Move `run_to_completion` utility to `crate::child` instead of `crate::commands::run`
This was really confusing as everything else in the `commands` module is a command
