+++ crates/uv/src/commands/mod.rs
@@ -72,7 +72,6 @@ mod project;
 mod publish;
 mod python;
 pub(crate) mod reporters;
-mod run;
 #[cfg(feature = "self-update")]
 mod self_update;
 mod tool;
